backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
SQLAlchemy>=1.4.0
pydantic>=1.10.0
python-dotenv>=1.0.0
openai>=1.0.0
httpx>=0.24.0
langchain>=0.0.200  # choose a reasonably recent LangChain release
tiktoken
alembic==1.11.1
fastapi
uvicorn[standard]
python-dotenv
openai
numpy
